+++ global.d.ts
@@ -4,6 +4,8 @@ import { GurpsActor } from './module/actor/actor.js'
 import { GurpsCombatant } from './module/combat/combatant.ts'
 import { GurpsItem } from './module/item.js'
 import { GurpsToken } from './module/token/gurps-token.ts'
+import { ManeuverDetail, ManeuverVisibility, RangeStrategy, RollBasedOnManeuverPolicy } from 'module/combat/types.ts'
+import { GurpsRange } from 'module/combat/ranges.js'
 
 export {}
 
@@ -46,6 +48,7 @@ declare global {
       currentSum(): number
       clear(): Promise<void>
       refreshPosition(): void
+      refresh(): void
     }
 
     DamageTables: {
@@ -61,9 +64,7 @@ declare global {
       getModifier(yards: number): number
     }
 
-    rangeObject: {
-      ranges: Array<{ modifier: number; max: number; penalty: number }>
-    }
+    rangeObject: GurpsRange
 
     Maneuvers: {
       get(id: string): { icon?: string } | undefined
@@ -205,16 +206,18 @@ declare global {
   }
 
   interface SettingConfig {
-    'gurps.rangeStrategy': 'Standard' | 'Simplified' | 'TenPenalties'
     'gurps.bucket-position': 'left' | 'right'
     'gurps.resource-tracker.manager': new (options?: any) => ResourceTracker.TemplateManager
     'gurps.resource-tracker.templates': Record<string, ResourceTrackerTemplate>
     'gurps.use-quick-rolls': AnyMutableObject
-    'gurps.use-on-target': boolean
-    'gurps.maneuver-visibility': string
-    'gurps.maneuver-detail': string
-    'gurps.maneuver-updates-move': boolean
-    'gurps.allow-roll-based-on-maneuver': string
+    'gurps.combat.rangeStrategy': RangeStrategy
+    'gurps.combat.use-on-target': boolean
+    'gurps.combat.maneuver-visibility': ManeuverVisibility
+    'gurps.combat.maneuver-detail': ManeuverDetail
+    'gurps.combat.maneuver-updates-move': boolean
+    'gurps.combat.allow-roll-based-on-maneuver': RollBasedOnManeuverPolicy
+    'gurps.combat.use-size-modifier-difference-in-melee': boolean
+    'gurps.combat.initiative-formula': String
     'gurps.show-confirmation-roll-dialog': boolean
     'gurps.modify-dice-plus-adds': boolean
     'gurps.pdf.basicset': String
@@ -224,7 +227,6 @@ declare global {
     'gurps.tracker-templates': new (options?: any) => Record<string, ResourceTrackerTemplate>
     'gurps.basicsetpdf': String
     'gurps.pdf-open-first': boolean
-    'gurps.use-size-modifier-difference-in-melee': boolean
     'gurps.portrait-hp-tinting': boolean
     'gurps.automatic-encumbrance': boolean
   }